dom/tests/mochitest/ajax/offline/test_badManifestMagic.html

<html xmlns="http://www.w3.org/1999/xhtml" manifest="http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/badManifestMagic.cacheManifest">
<head>
<title>bad manifest magic</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="/tests/dom/tests/mochitest/ajax/offline/offlineTests.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

var gGotChecking = false;

function finishTest() {
  OfflineTest.teardownAndFinish();
}

function handleError() {
  OfflineTest.ok(gGotChecking, "Expected checking event");
  OfflineTest.ok(true, "Expected error event");

  // These items are listed in the manifest, but the error should have
  // prevented them from being committed to the cache.
  var entries = [
    ["http://mochi.test:8888/tests/SimpleTest/SimpleTest.js", false],
    ["http://mochi.test:8888/tests/dom/tests/mochitest/ajax/offline/offlineTests.js", false]
  ];
  OfflineTest.checkCacheEntries(entries, finishTest);
}


if (OfflineTest.setup()) {
  // Don't expect a bunch of events.
  applicationCache.ondownloading = OfflineTest.failEvent;
  applicationCache.onupdateready = OfflineTest.failEvent;
  applicationCache.oncached = OfflineTest.failEvent;
  applicationCache.onnoupdate = OfflineTest.failEvent;
  applicationCache.onprogress = OfflineTest.failEvent;

  // ... but expect 'checking' and 'error'
  applicationCache.onchecking = function() { gGotChecking = true; };
  applicationCache.onerror = OfflineTest.priv(handleError);
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body>

</body>
</html>
